[#263] Extract approval history and DAP from FDA website
This code goes over each drug in the FDA website looking for its "Label and
Approval History" table. It then saves all rows, expanding the ones with
documents that link to a separate page (and not directly to a PDF). After this
is ran, we'll get a list of the approval history of all drugs with direct links
to the documents (e.g. labels, letters, reviews, ...) where they exist.

The item format is:

```javascript
{
   id: 'NDAXXXXXX-001',
   fda_application_num: 'NDAXXXXXX',
   action_date: Date(2012, 01, 12),
   supplement_number: 1,
   approval_type: 'Chemistry',
   documents: [
       {
           'name': 'Medical Review(s)',
           'urls': [
               'http://somewhere.com/part1.pdf',
               'http://somewhere.com/part2.pdf',
           ]
       },
       {
           'name': 'Label',
           'urls': [
               'http://somewhere.com/label.pdf',
           ]
       }
   ],
   notes: 'Label is not available on this site.'
}
```

All documents' URLs point directly to the PDF files. If a document is split in
multiple files (as sometimes Medical Reviews are), the order of the files in
`urls` is the same as they appeared on the page, so you could concatenate them
and get the full file.

I've also added the Betamax Python library, which allows recording and
replaying requests to third-party servers. This way our tests run against the
actual requests, but they're not dependent on the remote servers. If you'd like
to check if the scrapers are still working, you can simply remove the cassettes
at `tests/cassettes`.

Fixes opentrials/opentrials#263