[#658] Setup tox and disable FDADAP tests
The FDADAP tests were disabled because I had issues with Betamax recognizing the
current cassettes we have, and couldn't simply remove them as the site
changed (i.e. the tests break). We'll have to fix the collector in the future
anyway, but there's no need to do this now.

opentrials/opentrials#658