[#675] Stream NCT dump download
This lowers the memory usage but, more importantly, makes our code more
resilient to hiccups in the download of the data. The problem is that NCT is
generating the dump on the fly, so there can be some times where a downloaded
chunk is empty. If we're not streaming the download, requests end up throwing a
`requests.exceptions.ChunkedEncodingError` error.

Fixes opentrials/opentrials#675