[#678] Fix CrawlerProcess collectors and some minor adjustements to Sentry
The issue is that we were passing all constants defined in the
`collectors/base/config.py` file, which now include a Sentry client instance.
Scrapy tries to pickle the settings received and fails, because the Sentry
client instance isn't pickable.

The fix is to move the Scrapy settings to a separate constant and use that
explictly, so we only give Scrapy what is relevant to it.

I also did some adjustements to Sentry, specially for things to fail when an
exception is raised (instead of swallowing it).

Fixes opentrials/opentrials#678